Codify my GitHub icon and the brand coloring (#5)
Codify the icon and palette that I use when working with/as jrbeverly.

A while back I tried experimenting with a lot of project icons, but I stopped as it wasn't fitting exactly with what I wanted. I may explore that again but use the color palette defined here as the source definition for that.