Screenshot the pluralsight profile, and inline it on my website (#8)
Screenshot the pluralsight profile, inlining the contents of it on my website.

The screenshot doesn't offer interactive options, but that should be fine as all of the course certs are inlined into the directory. Any of the Skill IQ components (or in-progress courses) are no really needed for this.

The aim is to just capture any _useful_ pieces from the pluralsight profile before deletion.