Pluralsight component for referencing certificates (#11)
Inline pluralsight skill IQs and certificates of completion into my primary site.

Build the reference page with Hugo, then bundle it into my one-page static page. These components will need to be shuffled around so that it is a more cohesive structure.

The underlying way this is done in the repository will need to be cleaned up in the future. For now, this is sufficient as the primary aim of this is setting up the aesthetics of the page. Issues with fonts, css and dependencies can be resolved at a later date.