Stub out the sub-site for Udacity (#19)
Stub out a base site for Udacity, with the appropriate theme and missing content.

This is a stub put in place for udacity as adjustments are made to the structure of the websites. For now, its fine to copy the contents and make modifications as necessary.

I doubt the sites will remain as-is, but this needs to be brought in for other cleanup work.